requirement.txt
python-dotenv==0.19.2
numpy==1.22.2
scikit-learn==1.0.2
plotly==5.6.0
matplotlib==3.4.1
sty==1.0.4